Add surfing and splashy trie

+++ include/CompactTrie.hpp
@@ -14,7 +14,7 @@ public:
     // Currently, keys must be sorted
     explicit CompactTrie(std::vector<std::string> &keys);
     bool lookupPrefix(const std::string &prefix) override;
-    uint64_t getMemoryUsage() const override;
+    unsigned long long getMemoryUsage() const override;
     std::string getName() const override { return "BlindTrie"; }
 
 protected:
@@ -28,8 +28,8 @@ protected:
         virtual bool lookupNode(const std::string &key, uint64_t position);
         void insertKeys(uint64_t position, std::vector<std::string> &keys);
         void insertChildNode(char current, uint64_t position, std::vector<std::string> &keys);
-        virtual uint64_t getMemoryUsage() const;
-        uint64_t getChildrenMemoryUsage() const;
+        virtual unsigned long long getMemoryUsage() const;
+        unsigned long long getChildrenMemoryUsage() const;
 
         friend class CompactTrie;
     };
@@ -37,7 +37,7 @@ protected:
     class CompactNode : public TrieNode {
     protected:
         CompactNode(uint64_t fingerprint, uint64_t length);
-        uint64_t getMemoryUsage() const override;
+        unsigned long long getMemoryUsage() const override;
 
         uint8_t fingerprint_;
         uint64_t length_;
